# packages/backend/src/students/dto/batch-upsert-student.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { Type } from 'class-transformer';
import { IsArray, ValidateNested } from 'class-validator';
import { CreateStudentDto } from './create-student.dto';

export class BatchUpsertStudentDto {
  @ApiProperty({
    description: 'Array of students to create or update',
    type: [CreateStudentDto],
  })
  @IsArray()
  @ValidateNested({ each: true })
  @Type(() => CreateStudentDto)
  students: CreateStudentDto[];
}