# packages/frontend/backend_openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students": {
      "get": {
        "operationId": "StudentsController_getAll",
        "parameters": [
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all students or filtered by active status"
          }
        },
        "summary": "Get all students",
        "tags": [
          "students"
        ]
      },
      "post": {
        "operationId": "StudentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created"
          }
        },
        "summary": "Create new student",
        "tags": [
          "students"
        ]
      }
    },
    "/students/with-balance": {
      "get": {
        "operationId": "StudentsController_getAllWithBalance",
        "parameters": [
          {
            "name": "active",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all students with their financial balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentWithBalanceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all students with financial balance",
        "tags": [
          "students"
        ]
      }
    },
    "/students/search": {
      "get": {
        "operationId": "StudentsController_getByEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "description": "Student email",
            "schema": {
              "example": "student@example.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns student"
          },
          "404": {
            "description": "Student not found"
          }
        },
        "summary": "Get student by email",
        "tags": [
          "students"
        ]
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "StudentsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns student"
          }
        },
        "summary": "Get student by ID",
        "tags": [
          "students"
        ]
      },
      "patch": {
        "operationId": "StudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated"
          }
        },
        "summary": "Update student",
        "tags": [
          "students"
        ]
      },
      "delete": {
        "operationId": "StudentsController_deleteStudent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted"
          }
        },
        "summary": "Delete student",
        "tags": [
          "students"
        ]
      }
    },
    "/students/batch-upsert": {
      "post": {
        "operationId": "StudentsController_batchUpsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpsertStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Students created/updated. Returns count and the processed students."
          }
        },
        "summary": "Batch create or update students by email",
        "tags": [
          "students"
        ]
      }
    },
...
    "/payments/student/{studentId}": {
      "get": {
        "operationId": "PaymentsController_findByStudent",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payments for student",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Payment"
                  }
                }
              }
            }
          }
        },
        "summary": "Get payments by student id",
        "tags": [
          "payments"
        ]
      }
    },
...
    "/teachers/signup": {
      "post": {
        "operationId": "TeachersController_createTeacher",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teacher created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create new teacher",
        "tags": [
          "teachers"
        ]
      }
    },
    "/teachers/{id}": {
      "delete": {
        "operationId": "TeachersController_deleteTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Teacher ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher deleted successfully"
          }
        },
        "summary": "Delete teacher",
        "tags": [
          "teachers"
        ]
      }
    },
    "/teachers/batch-upsert": {
      "post": {
        "operationId": "TeachersController_batchUpsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpsertTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Teachers created/updated. Returns count and the processed teachers."
          }
        },
        "summary": "Batch create or update teachers by email",
        "tags": [
          "teachers"
        ]
      }
    },
...
    "/rooms": {
      "get": {
        "operationId": "RoomsController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns all rooms"
          }
        },
        "summary": "Get all rooms",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created"
          }
        },
        "summary": "Create new room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Room ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns room"
          }
        },
        "summary": "Get room by ID",
        "tags": [
          "rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Room ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated"
          }
        },
        "summary": "Update room",
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "RoomsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Room ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room deleted"
          }
        },
        "summary": "Delete room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/rooms/batch-upsert": {
      "post": {
        "operationId": "RoomsController_batchUpsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpsertRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rooms created/updated. Returns count and the processed rooms."
          }
        },
        "summary": "Batch create or update rooms by name",
        "tags": [
          "rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "KRUSANT API",
    "description": "API documentation for KRUSANT backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "StudentWithBalanceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "customRate": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "semester": {
            "type": "string"
          },
          "extraNotes": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "balance": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "customRate",
          "discount",
          "semester",
          "extraNotes",
          "active",
          "balance"
        ]
      },
      "PaymentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Payment date",
            "example": "2026-01-05"
          },
          "amount": {
            "type": "number",
            "description": "Payment amount",
            "example": 100
          },
          "comment": {
            "type": "string",
            "description": "Payment comment",
            "example": "Monthly payment"
          }
        },
        "required": [
          "date",
          "amount"
        ]
      },
      "ClassDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Class date",
            "example": "2026-01-05"
          },
          "cost": {
            "type": "number",
            "description": "Class cost",
            "example": 50
          },
          "type": {
            "type": "string",
            "description": "Class type",
            "example": "attended",
            "enum": [
              "attended",
              "missed"
            ]
          },
          "semester": {
            "type": "string",
            "description": "Semester",
            "example": "V"
          },
          "teacherId": {
            "type": "number",
            "description": "Teacher ID",
            "example": 1
          }
        },
        "required": [
          "date",
          "cost",
          "type",
          "semester",
          "teacherId"
        ]
      },
      "DebitDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Debit ID",
            "example": 1
          },
          "dueDate": {
            "type": "string",
            "description": "Due date",
            "example": "2026-01-31"
          },
          "amount": {
            "type": "number",
            "description": "Amount",
            "example": 100
          },
          "comment": {
            "type": "string",
            "description": "Comment",
            "example": "January fee"
          }
        },
        "required": [
          "id",
          "dueDate",
          "amount"
        ]
      },
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Student name",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Student email",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Student phone number",
            "example": "+1234567890"
          },
          "payments": {
            "description": "Payment records",
            "example": [
              {
                "date": "2026-01-05",
                "amount": 100,
                "comment": "Monthly payment"
              }
            ],
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          },
          "classes": {
            "description": "Class records",
            "example": [
              {
                "date": "2026-01-05",
                "cost": 50,
                "type": "attended",
                "semester": "V",
                "teacherId": 1
              }
            ],
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassDto"
            }
          },
          "debits": {
            "description": "Debit records",
            "example": [
              {
                "id": 1,
                "dueDate": "2026-01-31",
                "amount": 100,
                "comment": "January fee"
              }
            ],
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebitDto"
            }
          },
          "customRate": {
            "type": "number",
            "description": "Custom rate for this student",
            "example": 50
          },
          "discount": {
            "type": "number",
            "description": "Discount percentage",
            "example": 10
          },
          "semester": {
            "type": "string",
            "description": "Current semester",
            "example": "V"
          },
          "extraNotes": {
            "type": "string",
            "description": "Additional notes",
            "example": "",
            "default": ""
          },
          "active": {
            "type": "boolean",
            "description": "Student active status",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name",
          "email",
          "payments",
          "classes",
          "semester",
          "extraNotes",
          "active"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Student name",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Student email",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Student phone number",
            "example": "+1234567890"
          },
          "payments": {
            "description": "Payment records",
            "example": [
              {
                "date": "2026-01-05",
                "amount": 100,
                "comment": "Monthly payment"
              }
            ],
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          },
          "classes": {
            "description": "Class records",
            "example": [
              {
                "date": "2026-01-05",
                "cost": 50,
                "type": "attended",
                "semester": "V",
                "teacherId": 1
              }
            ],
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassDto"
            }
          },
          "debits": {
            "description": "Debit records",
            "example": [
              {
                "id": 1,
                "dueDate": "2026-01-31",
                "amount": 100,
                "comment": "January fee"
              }
            ],
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DebitDto"
            }
          },
          "customRate": {
            "type": "number",
            "description": "Custom rate for this student",
            "example": 50
          },
          "discount": {
            "type": "number",
            "description": "Discount percentage",
            "example": 10
          },
          "semester": {
            "type": "string",
            "description": "Current semester",
            "example": "V"
          },
          "extraNotes": {
            "type": "string",
            "description": "Additional notes",
            "example": "",
            "default": ""
          },
          "active": {
            "type": "boolean",
            "description": "Student active status",
            "example": true,
            "default": true
          }
        }
      },
      "BatchUpsertStudentDto": {
        "type": "object",
        "properties": {
          "students": {
            "description": "Array of students to create or update",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateStudentDto"
            }
          }
        },
        "required": [
          "students"
        ]
      },
...
      "Payment": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTeacherDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Teacher name",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Teacher email",
            "example": "teacher@example.com"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "BatchUpsertTeacherDto": {
        "type": "object",
        "properties": {
          "teachers": {
            "description": "Array of teachers to create or update",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTeacherDto"
            }
          }
        },
        "required": [
          "teachers"
        ]
      },
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Room name",
            "example": "Room A"
          },
          "capacity": {
            "type": "number",
            "description": "Capacity of the room",
            "example": 20
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Room name",
            "example": "Room A"
          },
          "capacity": {
            "type": "number",
            "description": "Capacity of the room",
            "example": 20
          }
        }
      },
      "BatchUpsertRoomDto": {
        "type": "object",
        "properties": {
          "rooms": {
            "description": "Array of rooms to create or update",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateRoomDto"
            }
          }
        },
        "required": [
          "rooms"
        ]
      },
...
    }
  }
}